backend/requirements.txt
# ── Core Framework ──
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# ── Database (SQLite via aiosqlite) ──
sqlalchemy==2.0.36
aiosqlite==0.20.0
greenlet==3.3.2

# ── Graph Engine ──
networkx==3.4.2

# ── Data Ingestion ──
pandas==2.2.3

# ── Utilities ──
python-dotenv==1.0.1
structlog==24.4.0

# ── Dev & Testing ──
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1

# ── LLM Integration ──
groq

